// web/src/api/generated/models/PlanDetail.ts
export interface PlanDetail {
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof PlanDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof PlanDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof PlanDetail
     */
    createdBy: string;
    /**
     * 
     * @type {AuthorSummary}
     * @memberof PlanDetail
     */
    creator: AuthorSummary;
    /**
     * 
     * @type {string}
     * @memberof PlanDetail
     */
    description: string | null;
    /**
     * 
     * @type {Date}
     * @memberof PlanDetail
     */
    experiencedOn: Date | null;
    /**
     * 
     * @type {string}
     * @memberof PlanDetail
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof PlanDetail
     */
    placeId: string | null;
    /**
     * 
     * @type {Date}
     * @memberof PlanDetail
     */
    plannedEnd: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof PlanDetail
     */
    plannedOn?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof PlanDetail
     */
    plannedStart: Date | null;
    /**
     * 
     * @type {string}
     * @memberof PlanDetail
     */
    sourceWishId: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlanDetail
     */
    spaceId: string;
    /**
     * 
     * @type {PlanStatus}
     * @memberof PlanDetail
     */
    status: PlanStatus;
    /**
     * 
     * @type {string}
     * @memberof PlanDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof PlanDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof PlanDetail
     */
    version: number;
}